+++ app/Http/Controllers/ApplicationsController.php
@@ -9,14 +9,12 @@
  */
 namespace App\Http\Controllers;
 
+use App\Basket\Application;
 use App\Basket\Installation;
 use App\Exceptions\RedirectException;
-use App\Http\Requests;
-use App\Basket\Application;
 use Illuminate\Http\Request;
-use Illuminate\Support\Collection;
-use PayBreak\Sdk\Gateways\ApplicationGateway;
 use Illuminate\Support\Facades\Log;
+use PayBreak\Sdk\Gateways\ApplicationGateway;
 
 /**
  * Class ApplicationsController
@@ -46,7 +44,7 @@ class ApplicationsController extends Controller
      *
      * @author WN, MS
      * @param int $installation
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function index($installation)
     {
@@ -68,6 +66,7 @@ class ApplicationsController extends Controller
             [
                 'default_dates' => $filterDates,
                 'ext_current_status' => $this->fetchFilterValues($applications, 'ext_current_status'),
+                'ext_finance_option_group' => $this->fetchFilterValues($applications, 'ext_finance_option_group'),
             ]
         );
     }
@@ -77,7 +76,7 @@ class ApplicationsController extends Controller
      *
      * @param int $installation
      * @param  int $id
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function show($installation, $id)
     {
@@ -99,7 +98,7 @@ class ApplicationsController extends Controller
      *
      * @param $installation
      * @param  int $id
-     * @return Response
+     * @return \Illuminate\View\View
      */
     public function edit($installation, $id)
     {
@@ -116,7 +115,7 @@ class ApplicationsController extends Controller
      * @param $installation
      * @param int $id
      * @param Request $request
-     * @return Response
+     * @return \Illuminate\Http\RedirectResponse
      * @throws RedirectException
      */
     public function update($installation, $id, Request $request)
@@ -146,7 +145,7 @@ class ApplicationsController extends Controller
      * @param $installation
      * @param int $id
      * @return \Illuminate\Http\RedirectResponse
-     * @throws ApplicationsController
+     * @throws RedirectException
      */
     public function fulfil($installation, $id)
     {
@@ -179,11 +178,11 @@ class ApplicationsController extends Controller
 
     /**
      * @author WN
-     * @param int $installation
-     * @param int $id
+     * @param $installation
+     * @param $id
      * @param Request $request
      * @return \Illuminate\Http\RedirectResponse
-     * @throws ApplicationsController
+     * @throws RedirectException
      */
     public function requestCancellation($installation, $id, Request $request)
     {
@@ -202,36 +201,26 @@ class ApplicationsController extends Controller
         );
     }
 
-    /**
-     * Display pending cancellation list.
+    /**Display pending cancellation list.
      *
-     * @author SD
+     * @author SD, EB
+     * @param $installationId
      * @return \Illuminate\View\View
-     * @throws \App\Exceptions\RedirectException
      */
     public function pendingCancellations($installationId)
     {
-        $installation = $this->fetchModelByIdWithMerchantLimit((new Installation()), $installationId, 'installation', '/');
+        $installation = Installation::query()->findOrFail($installationId);
 
-        $pendingCancellations = Collection::make(
-            $this
-                ->applicationGateway
-                ->getPendingCancellations($installation->ext_id, $installation->merchant->token)
-        );
+        $pendingCancellations = Application::query()
+            ->where('installation_id', '=', $installationId)
+            ->where('ext_current_status', '=', 'pending_cancellation')
+            ->get();
 
-        $local = [];
-
-        // Shouldn't need to do this but leaving for refactoring as this
-        // is done across all code base
-        foreach ($pendingCancellations as $key => $pendingCancellation) {
-            $pendingCancellations[$key] = (object) $pendingCancellation;
-            $local[$pendingCancellation['id']] = Application::where('ext_id', '=', $pendingCancellation['id'])->first();
-        }
-
-        return View('applications.pending-cancellation', [
-            'applications' => $pendingCancellations,
-            'local' => $local,
-        ]);
+        return View('applications.pending-cancellation',
+            [
+                'applications' => $pendingCancellations,
+                'installation' => $installation,
+            ]);
     }
 
     /**
@@ -337,7 +326,7 @@ class ApplicationsController extends Controller
 
     /**
      * @author WN
-     * @param $action
+     * @param string $action
      * @param $id
      * @return \Illuminate\View\View
      * @throws RedirectException