+++ database/migrations/2015_09_30_142336_add_active_product_group_to_applications_table.php
@@ -1,0 +1,37 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+/**
+ * Add finance product group to table
+ *
+ * @author SL
+ * Class AddActiveProductGroupToApplicationsTable
+ */
+class AddActiveProductGroupToApplicationsTable extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->string('ext_finance_option_group')->nullable()->after('ext_order_validity');
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('applications', function (Blueprint $table) {
+            $table->dropColumn('ext_finance_option_group');
+        });
+    }
+}