+++ resources/views/account/show.blade.php
@@ -12,9 +12,9 @@
 
     @include('includes.page.breadcrumb')
     <div class="panel panel-default">
-        <div class="panel-heading"><h3 class="panel-title">USER DETAILS</h3></div>
+        <div class="panel-heading"><h3 class="panel-title">User Details</h3></div>
         <div class="panel-body">
-            {!! Form::model($user) !!}
+            {!! Form::model($user, ['class' => 'form-horizontal']) !!}
             <div class="row">
                 <div class="col-xs-2 col-sm-4 col-md-2 col-lg-2">
                     <div class="thumbnail">
@@ -22,13 +22,18 @@
                     </div>
                 </div>
                 <div class="col-xs-10 col-sm-8 col-md-10 col-lg-10">
+                    &nbsp;
                     <div class="form-group">
-                        {!! Form::label('name', 'Name: ', ['class' => 'col-sm-3 control-label']) !!}
-                        {!! Form::text('name', null, ['class' => 'form-control', 'readonly']) !!}
+                        {!! Form::label('name', 'Name: ', ['class' => 'col-sm-2 control-label']) !!}
+                        <div class="col-sm-10">
+                            {!! Form::text('name', null, ['class' => 'form-control']) !!}
+                        </div>
                     </div>
                     <div class="form-group">
-                        {!! Form::label('email', 'Email: ', ['class' => 'col-sm-3 control-label']) !!}
-                        {!! Form::text('email', null, ['class' => 'form-control', 'readonly']) !!}
+                        {!! Form::label('email', 'Email: ', ['class' => 'col-sm-2 control-label']) !!}
+                        <div class="col-sm-10">
+                            {!! Form::text('email', null, ['class' => 'form-control']) !!}
+                        </div>
                     </div>
                 </div>
             </div>