+++ resources/views/applications/pending-cancellation.blade.php
@@ -3,7 +3,7 @@
 @section('content')
 
     <h1>Pending Cancellations</h1>
-    @include('includes.page.breadcrumb', ['over' => [1 => isset($applications[0]->installation) ? $applications[0]->installation : Request::segment(2)], 'permission' => [0 => Auth::user()->can('merchants-view'), 1 => Auth::user()->can('merchants-view')]])
+    @include('includes.page.breadcrumb', ['over' => [1 => $installation->name], 'permission' => [0 => Auth::user()->can('merchants-view'), 1 => Auth::user()->can('merchants-view')]])
 
     {!! Form::open(array('url' => Request::url() . '/?' . Request::server('QUERY_STRING'), 'method' => 'get',  'onsubmit'=>"return submitFilter()")) !!}
 
@@ -12,24 +12,22 @@
             {{-- TITLES --}}
             <th>Retailer Ref.</th>
             <th>Name</th>
-            {{--<th>Cancelled Reason</th>--}}
-            {{--<th>Requested</th>--}}
+            <th>Cancellation Fee</th>
+            <th>Cancelled Reason</th>
+            <th>Requested</th>
             <th></th>
         </tr>
     @forelse($applications as $item)
         <tr>
-            <td>{{ $item->order['reference'] }}</td>
-            <td>{{ trim($item->customer['title'] . ' ' . $item->customer['first_name'] . ' '.  $item->customer['last_name']) }}</td>
-            {{--<td>{{ $item->cancellation['description'] }}</td>--}}
-            {{--<td>{{ date('d/m/Y H:i', strtotime($item->cancellation['requested_date'])) }}</td>--}}
+            <td>{{ $item->ext_order_reference }}</td>
+            <td>{{ trim($item->ext_customer_title . ' ' . $item->ext_customer_first_name . ' '.  $item->ext_customer_last_name) }}</td>
+            <td>{{ '&pound;' . number_format($item->ext_cancellation_fee_amount/100, 2) }}</td>
+            <td>{{ $item->ext_cancellation_description }}</td>
+            <td>{{ date('d/m/Y H:i', strtotime($item->ext_cancellation_requested_date)) }}</td>
 
             <td class="text-right">
                 <div class="btn-group">
-                    @if($local[$item->id])
-                    <a href="/installations/{{$local[$item->id]['installation']->id}}/applications/{{$local[$item->id]['id']}}" type="button" class="btn btn-default btn-xs"> View </a>
-                    @else
-                    <button type="button" class="btn btn-default dropdown-toggle btn-xs @if(empty($local[$item->id])) disabled @endif" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"> View </button>
-                    @endif
+                    <a href="/installations/{{$item->installation->id}}/applications/{{$item->id}}" type="button" class="btn btn-default btn-xs"> View </a>
                 </div>
             </td>
         </tr>