+++ resources/views/installations/show.blade.php
@@ -13,6 +13,7 @@
 
     <ul class="nav nav-tabs">
         <li class="active"><a data-toggle="tab" href="#part1">Installation Details</a></li>
+        <li><a data-toggle="tab" href="#part2">Finance Product Details</a></li>
     </ul>
     <div class="tab-content">
         <div id="part1" class="tab-pane fade in active">
@@ -94,6 +95,85 @@
             @endif
 
         </div>
-
+        <div id="part2" class="tab-pane fade in">
+            <br/>
+            <table class="table table-bordered">
+                <thead>
+                    <tr>
+                        <th colspan="2">Product</th>
+                        <th colspan="4">Terms</th>
+                        <th colspan="2">Order Range</th>
+                        <th colspan="4">Deposit</th>
+                        <th colspan="4">Merchant Fees</th>
+                    </tr>
+                    <tr>
+                        <th>Product ID</th>
+                        <th>Full Product Name</th>
+                        <th>Holiday</th>
+                        <th>Payments</th>
+                        <th>Interest</th>
+                        <th>Fee</th>
+                        <th>Min</th>
+                        <th>Max</th>
+                        <th>Min %</th>
+                        <th>Max %</th>
+                        <th>Min £</th>
+                        <th>Max £</th>
+                        <th>%</th>
+                        <th>Min £</th>
+                        <th>Max £</th>
+                        <th>Cancellation</th>
+                    </tr>
+                </thead>
+                <tbody>
+                    @forelse($products as $group)
+                        <tr class="clickable" data-toggle="collapse" id="{{$group->getId()}}" data-target=".{{$group->getId()}}collapsed">
+                            <td colspan="16"><span class="glyphicon glyphicon-plus"></span><span class="glyphicon glyphicon-minus hidden"></span> {{$group->getName()}}</td>
+                        </tr>
+                        @forelse($group->getProducts() as $productData)
+                            <tr class="collapse out {{$group->getId()}}collapsed">
+                                <td><code>{{$productData->getId()}}</code></td>
+                                <td>{{$productData->getName()}}</td>
+                                <td>{{$productData->getHolidays()}}</td>
+                                <td>{{$productData->getPayments()}}</td>
+                                <td>{{number_format($productData->getPerAnnumInterestRate(), 1)}}%</td>
+                                <td>{{'&pound;' . number_format($productData->getCustomerServiceFee()/100, 2)}}</td>
+                                <td>{{'&pound;' . number_format($productData->getOrder()->getMinimumAmount()/100, 2)}}</td>
+                                <td>{{'&pound;' . number_format($productData->getOrder()->getMaximumAmount()/100, 2)}}</td>
+                                <td>{{number_format($productData->getDeposit()->getMinimumPercentage(), 2)}}%</td>
+                                <td>{{number_format($productData->getDeposit()->getMaximumPercentage(), 2)}}%</td>
+                                <td>{{'&pound;' . number_format($productData->getDeposit()->getMinimumAmount()/100, 2)}}</td>
+                                <td>{{'&pound;' . number_format($productData->getDeposit()->getMaximumAmount()/100, 2)}}</td>
+                                <td>{{number_format($productData->getMerchantFees()->getPercentage(), 2)}}%</td>
+                                <td>{{'&pound;' . number_format($productData->getMerchantFees()->getMinimumAmount()/100, 2)}}</td>
+                                <td>{{'&pound;' . number_format($productData->getMerchantFees()->getMaximumAmount()/100, 2)}}</td>
+                                <td>{{'&pound;' . number_format($productData->getMerchantFees()->getCancellation()/100, 2)}}</td>
+                            <tr>
+                        @empty
+                            <tr>
+                                <td>There are no products for {{$group->getName()}}</td>
+                            <tr>
+                        @endforelse
+                    @empty
+                        <tr>There are no products to display</tr>
+                    @endforelse
+                </tbody>
+            </table>
+        </div>
     </div>
+@endsection
+
+@section('scripts')
+    <script>
+        $('tr.clickable').on('click', function () {
+            if()
+            $(this).find('span').each(function() {
+                if($(this).hasClass('hidden')) {
+                    $(this).removeClass('hidden');
+                } else {
+                    $(this).addClass('hidden');
+                }
+            });
+        });
+    </script>
 @endsection